pyproject.toml
[tool.poetry]
name = "text-adventure"
version = "0.1.0"
description = "A text-based roguelike including dungeons, monsters and magic 🎉"
authors = ["Alexis Moins <moinsalexis@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
colorama = "^0.4.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
autopep8 = "^1.6.0"
pre-commit = "^2.18.1"

[tool.autopep8]
max_line_length = 120

[tool.pytest.ini_options]
pythonpath = ["."]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
